% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DengueForecast.R
\name{DengueFC}
\alias{DengueFC}
\title{Dengue Forecast}
\usage{
DengueFC(h)
}
\arguments{
\item{h}{forecast horizon in weeks (a number greater than 0)}
}
\examples{
DengueFC(10) # Returns forecasts for the next 10 weeks
}
